[fe] refactor (#19) 컴포넌트 분류 - 컴포넌트를 페이지 별로 분리
