+++ client/src/components/driverSignup/DriverSignupForm.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { Button } from 'antd-mobile';
 import { withFormik, FormikProps } from 'formik';
-import SignupLabelInput from './SignupLabelInput';
-import driverSignupValidation from '../utils/driverSignupValidation';
-import { driverSignupFormValues } from '../types';
+import SignupLabelInput from '../SignupLabelInput';
+import driverSignupValidation from '../../utils/driverSignupValidation';
+import { driverSignupFormValues } from '../../types';
 
 const InnerForm: React.ElementType = ({
   values,
